Introduce new scheme of extension related predefined macro
The motivation of this scheme is have an unify naming scheme for
extension macro and add the capability to checking version.

The new naming scheme is `__riscv_<extension name>`, and it will define
the value to `<MAJOR_VERSION> * 1,000,000 + <MINOR_VERSION> * 1,000 + <REVISION_VERSION>`

For example:
 - F-extension v2.2 will define `__riscv_f` as `2002000`.
 - B-extension v0.92 will define `__riscv_b` as `92000`.

We also introduce a new macro `__riscv_arch_test_marco` to check this
test marco set is available.