Add Defined misa Bits
This adds the bits that are defined already in `misa.Extensions`, in
the privileged spec, following the note above the table, and existing
ratified extensions.

While we would not expect all of these to be present in an OS-level
hwcap interface (for example the H bit), I think it is fine to allocate
these for the single-letter part of the bitmask, so these bits are not
otherwise used.